generation time for pdbs is now double instead of int, so we can set limits bellow one sec. Float is not available as a data type in get_opts, so I used double instead, either should work fine.

+++ src/search/pdbs/pattern_collection_generator_manual_multiple.h
@@ -15,7 +15,7 @@ namespace pdbs {
 class PatternCollectionGeneratorManualMultiple : public PatternCollectionGenerator {
     std::shared_ptr<PatternCollection> patterns;
     std::shared_ptr<PDBFactory> pdb_type;
-    int time_limit;
+    double time_limit;
 
 public:
     explicit PatternCollectionGeneratorManualMultiple(const options::Options &opts);
