+++ src/search/pdbs/pdb_factory_explicit.h
@@ -27,7 +27,7 @@ namespace pdbs {
 	create_pdb(const TaskProxy & task, 
 		   const Pattern &pattern, 
 		   const std::vector<int> &operator_costs = std::vector<int>(), 
-		   int time_limit = std::numeric_limits<int>::max()
+		   double time_limit = std::numeric_limits<int>::max()
 	    );
 
     public:
